Move `WebSocketClientMock` to separate file
